Add tests for limit clause and fix issues
