--- tests/jballerina-unit-test/src/test/resources/test-src/query/onConflictClause-limitClause.bal
@@ -1,69 +1,0 @@
-type Person record {|
-    string firstName;
-    string lastName;
-    int age;
-|};
-
-type FullName record{|
-	string firstName;
-	string lastName;
-|};
-
-function testOnConflictClause() returns Person[]|error {
-   
-    Person p1 = {firstName: "Alex", lastName: "George", age: 23};
-    Person p2 = {firstName: "Ranjan", lastName: "Fonseka", age: 30};
-    Person p3 = {firstName: "John", lastName: "David", age: 33};
-
-    Person[] personList = [p1, p2, p3];
-
-    Person[]|error outputPersonList =
-            from var person in personList
-            select {
-                   firstName: person.firstName,
-                   lastName: person.lastName,
-                   age: person.age
-            }
-            on conflict error("PersonNotFound");
-
-    return outputPersonList;
-}
-
-function testLimitClause() returns Person[] {
-
-    Person p1 = {firstName: "Alex", lastName: "George", age: 23};
-    Person p2 = {firstName: "Ranjan", lastName: "Fonseka", age: 30};
-    Person p3 = {firstName: "John", lastName: "David", age: 33};
-
-    Person[] personList = [p1, p2, p3];
-
-    Person[] outputPersonList =
-            from var person in personList
-            select {
-                   firstName: person.firstName,
-                   lastName: person.lastName,
-                   age: person.age
-            }
-            limit 2;
-
-    return outputPersonList;
-}
-
-function testLimitClauseWithQueryAction() returns FullName[]{
-
-    Person p1 = {firstName: "Alex", lastName: "George", age: 23};
-    Person p2 = {firstName: "Ranjan", lastName: "Fonseka", age: 30};
-    Person p3 = {firstName: "John", lastName: "David", age: 33};
-
-    Person[] personList = [p1, p2, p3];
-    FullName[] nameList = [];
-
-    var x =  from var person in personList
-            do {
-                FullName fullName = {firstName: person.firstName, lastName: person.lastName};
-                nameList[nameList.length()] = fullName;
-            }
-            limit 2;
-
-    return nameList;
-}
